flickr download and display in Collection View working somewhat well.  also does prefetch and downloads again in PhotoAlbumViewController in case the download was interrupted/the app was terminated early during the download.  care is made to ensure that there are not 2 download sessions in progress at the same time.

+++ Virtual Tourist/PhotoAlbumViewController.swift
@@ -13,10 +13,12 @@ import CoreData
 
 class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDelegate, UICollectionViewDelegate, UICollectionViewDataSource {
     var pin : Pin!
+    var fullyDownloaded = false
     
     @IBOutlet weak var mapView: MKMapView!
     @IBOutlet weak var collectionView: UICollectionView!
     @IBOutlet weak var newCollectionButton: UIButton!
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         println("pin coordinate=(\(pin!.latitude),\(pin!.longitude)")
@@ -26,6 +28,30 @@ class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDel
         var error: NSError? = nil
         
         fetchedResultsController.performFetch(&error)
+        let sectionInfo = self.fetchedResultsController.sections![0] as! NSFetchedResultsSectionInfo
+        println("collectionView numberOfItemsInSection=\(sectionInfo.numberOfObjects)")
+        for (var i = 0; i < sectionInfo.numberOfObjects; i++) {
+            if let photo = self.fetchedResultsController.objectAtIndexPath(NSIndexPath(forRow: i, inSection: 0)) as? Photo {
+                print("it's a photo")
+                if let localPath = photo.localPath {
+                    print(" with localPath=\(localPath)")
+                } else {
+                    print(" without localPath")
+                }
+                
+                if let pin = photo.pin {
+                    print(" with pin=\(pin.objectID)")
+                } else {
+                    print(" without pin")
+                }
+                
+                println("")
+                
+            } else {
+                println("not a photo")
+            }
+        }
+        
         
         if let error = error {
             println("Unresolved error \(error), \(error.userInfo)")
@@ -40,8 +66,13 @@ class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDel
             //let pin = object as! Pin
             //addPinToMapWithPin(pin)
         }
+        collectionView.delegate = self
+        collectionView.dataSource = self
+        //collectionView.reloadData()
+        collectionView.alwaysBounceVertical = true
+        collectionView.scrollEnabled = true
         
-        
+        PhotoDownloader.sharedInstance().downloadPhotos(pin)
         println("viewDidLoad complete")
 
         
@@ -73,6 +104,10 @@ class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDel
     // open detail view
     func collectionView(collectionView: UICollectionView, didSelectItemAtIndexPath indexPath: NSIndexPath) {
         // delete Photo, only allow delete when all photos are downloaded
+        if !fullyDownloaded {
+            return
+        }
+        
         if let photo = self.fetchedResultsController.objectAtIndexPath(indexPath) as? Photo {
             sharedContext.deleteObject(photo)
             CoreDataStackManager.sharedInstance().saveContext()
@@ -82,22 +117,33 @@ class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDel
     
     func collectionView(collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         let sectionInfo = self.fetchedResultsController.sections![section] as! NSFetchedResultsSectionInfo
+        println("collectionView numberOfItemsInSection=\(sectionInfo.numberOfObjects)")
         return sectionInfo.numberOfObjects
     }
     
     // The cell that is returned must be retrieved from a call to -dequeueReusableCellWithReuseIdentifier:forIndexPath:
     func collectionView(collectionView: UICollectionView, cellForItemAtIndexPath indexPath: NSIndexPath) -> UICollectionViewCell {
+        println("collectionView cellForItemAtIndexPath=\(indexPath.row)")
         let cell = collectionView.dequeueReusableCellWithReuseIdentifier("CellIdentifier", forIndexPath: indexPath) as! UICollectionViewCell
         if let photo = self.fetchedResultsController.objectAtIndexPath(indexPath) as? Photo {
+            let imageView = cell.viewWithTag(1001) as! UIImageView
             if let localPath = photo.localPath {
-                let imageView = cell.viewWithTag(1001) as! UIImageView
-                imageView.image = UIImage(contentsOfFile: localPath)
+                if NSFileManager.defaultManager().fileExistsAtPath(localPath) {
+                    imageView.image = UIImage(contentsOfFile: localPath)
+                    println("using actual image")
+                } else {
+                    println("actual image is broken")
+                }
+            } else {
+                //imageView.image = UIImage(named: "placeHolder")
+                println("using default image")
             }
         }
         return cell
     }
     
     func numberOfSectionsInCollectionView(collectionView: UICollectionView) -> Int {
+        println("numberOfSectionsInCollectionView=1")
         return 1
     }
     
@@ -107,12 +153,7 @@ class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDel
     
     func sanitizePhoto(anObject : AnyObject?) -> Photo? {
         if let photo = anObject as? Photo {
-            // only update for photos with same pin, could be other background operations going on
-            if photo.pin!.id == pin.id {
-                return photo
-            } else {
-                println("sanitizePhoto photo is not the same as current Pin")
-            }
+            return photo
         } else {
             println("sanitizePhoto object not a Photo")
         }
@@ -145,16 +186,20 @@ class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDel
                 case .Delete:
                     // TODO: remove picture from CollectionView
                     collectionView.deleteItemsAtIndexPaths([indexPath])
+                    println("deleted image at indexPath.row=\(indexPath.row)")
                     break
                 case .Insert:
                     // TODO: add picture to CollectionView if ready, otherwise, placeholder
                     collectionView.insertItemsAtIndexPaths([indexPath])
+                    println("inserted image at indexPath.row=\(indexPath.row)")
                     break
                 case .Update:
                     // TODO: add picture if changed from not ready to ready (i.e. path exists)
                     collectionView.reloadItemsAtIndexPaths([indexPath])
+                    println("updated image at indexPath.row=\(indexPath.row)")
                     return
                 default:
+                    println("default")
                     return
                 }
             } else {
@@ -178,6 +223,8 @@ class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDel
         
         // Create the fetch request
         let fetchRequest = NSFetchRequest(entityName: "Photo")
+
+        fetchRequest.predicate = NSPredicate(format: " pin == %@", self.pin)
         
         // Add a sort descriptor.
         fetchRequest.sortDescriptors = [NSSortDescriptor(key: "id", ascending: false)]
@@ -185,11 +232,20 @@ class PhotoAlbumViewController : UIViewController, NSFetchedResultsControllerDel
         // Create the Fetched Results Controller
         let fetchedResultsController = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: self.sharedContext, sectionNameKeyPath: nil, cacheName: nil)
         
+        //fetchedResultsController.
+        //fetchedResultsController.
+        
         // Return the fetched results controller. It will be the value of the lazy variable
         
         return fetchedResultsController
         } ()
     
     @IBAction func newCollection(sender: AnyObject) {
+    }
+    
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        println("didReceiveMemoryWarning")
+        // Dispose of any resources that can be recreated.
     }
 }
